+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `.` tilecode now shows a more obvious marker since "." was mostly invisible
   - Fixed temple `air` chunks which has accidentally been a copy/paste of `temple` chunk
   - Fixed typo in Haunted Castle rooms that said "Caster" instead of "Castle"
+  - Use closed door for Mothership, Ice Caves, Wet Fur, and Tutorial Entrance
 
 ### Added